# spec/openapi.yaml
openapi: 3.1.0
info:
  title: Takomo
  version: "1.0.0-draft1"
  description: |
    Central task store for agent fleets. One authority, bearer-token auth,
    hierarchical tickets, enforced per-project workflows, atomic claim/lease,
    append-only event log.

    Design rules this contract encodes:
    - Most writes are commutative (comments, field sets, metadata merge); CAS
      (`If-Match` on `version`) is only needed for whole-body replacement.
    - Every rejection teaches: error bodies state what is allowed and how to
      proceed (`allowed_transitions`, `remedy`).
    - Claiming is one atomic operation (`POST /ready/claim`), never read-then-claim.
    - The event log (`global_seq`) is the only push/poll surface; all delivery
      mechanisms (SSE, hooks, channels) layer on the `since` cursor.
    - `/v1` is frozen: additive evolution only; experiments live in `metadata`.
servers:
  - url: /v1
security:
  - bearer: []
...
paths:
...
  /projects/{project}/workflow:
    get:
      tags: [workflow]
      summary: Get the project's workflow definition
      parameters: [ { $ref: "#/components/parameters/project" } ]
      responses:
        "200":
          description: The active workflow (see workflow-format.md)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workflow" }
    put:
      tags: [workflow]
      summary: Replace the project's workflow (admin scope; must remain valid for existing tickets)
      parameters: [ { $ref: "#/components/parameters/project" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Workflow" }
      responses:
        "200": { description: Replaced }
        "422":
          description: Invalid workflow (unreachable states, tickets stranded)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: See auth.md — `tk_`-prefixed tokens, scoped, hashed at rest.
    shareBearer:
      type: http
      scheme: bearer
      description: |
        A SHARE token (`tks_`-prefixed), minted by `POST /v1/shares` and hashed
        at rest like a normal token. It is a distinct auth path: it authorizes
        ONLY the read-only `/v1/shares/self*` endpoints, bounded to its scope,
        and is rejected on every normal endpoint. Expired/revoked shares return
        `410 Gone`. See auth.md.
  parameters:
...
    project:
      name: project
      in: path
      required: true
      schema: { type: string }
...
  schemas:
...
    Workflow:
      type: object
      description: See workflow-format.md for authoring format and semantics
      required: [name, initial, states, transitions]
      properties:
        name: { type: string }
        initial: { type: string }
        states:
          type: array
          items:
            type: object
            required: [id, category]
            properties:
              id: { type: string }
              category: { type: string, enum: [todo, in_progress, blocked, review, done, cancelled] }
              claimable: { type: boolean, default: false }
              terminal: { type: boolean, default: false }
        transitions:
          type: array
          items:
            type: object
            required: [from, to]
            properties:
              from: { type: string }
              to: { type: string }
              requires:
                type: array
                items: { type: string }
                description: "any of: claim | scope:<scope> | guard:<guard-id>"
        guards:
          type: object
          description: named guard configs, e.g. no_open_children on done-category entry
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string, description: "stable machine code, e.g. conflict.version, claim.held, fence.stale (superseded lease), fence.invalid (fence never issued — a client bug)" }
        message: { type: string, description: written for LLM readers — states what is wrong AND what to do }
        remedy: { type: string, description: the exact next call to make, when one exists }
        current_version: { type: integer }
        details: { type: object }
...